parking_app/requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.2.1
Django==5.2
djangorestframework==3.16.0
executing==2.2.0
idna==3.10
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.5.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
